# enh: revert back to array for llTable/mlTable/ofTable