--- a/zstd/seqdec.go
+++ b/zstd/seqdec.go
@@ -97,7 +97,7 @@ func (s *sequenceDecs) initialize(br *bitReader, hist *history, literals, out []
 func (s *sequenceDecs) decode(seqs int, br *bitReader, hist []byte) error {
 	startSize := len(s.out)
 	// Go 1.17+ no longer bounds checks access to array; avoid slice indirection overhead.
-	llTable, mlTable, ofTable := s.litLengths.fse.dt, s.matchLengths.fse.dt, s.offsets.fse.dt
+	llTable, mlTable, ofTable := &s.litLengths.fse.dt, &s.matchLengths.fse.dt, &s.offsets.fse.dt
 	llState, mlState, ofState := s.litLengths.state.state, s.matchLengths.state.state, s.offsets.state.state
 
 	for i := seqs - 1; i >= 0; i-- {