The extra indirection of creating a slice with size maxTableSize is no longer needed; this results in a 25% performance increase on x86_64 for some workloads (esp html_x_4), and flat performance on aarch64.

go1.17/amd64:
```
benchmark                                                    old ns/op     new ns/op     delta
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-12               3494468       3379259       -3.30%
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-12           687076        681387        -0.83%
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-12            11085443      10924817      -1.45%
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-12              8407670       8158085       -2.97%
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-12            2825509       2749297       -2.70%
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-12             3736775       3604104       -3.55%
BenchmarkDecoder_DecoderSmall/html_x_4.zst-12                1903319       1417892       -25.50%
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-12          159818        154350        -3.42%
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-12          75035         75623         +0.78%
BenchmarkDecoder_DecoderSmall/urls.10K.zst-12                9421182       9058897       -3.85%
BenchmarkDecoder_DecoderSmall/html.zst-12                    795930        771416        -3.08%
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-12           66842         65826         -1.52%
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-12                  435892        424669        -2.57%
BenchmarkDecoder_DecodeAll/geo.protodata.zst-12              84058         90552         +7.73%
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-12               1383197       1352505       -2.22%
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-12                 1042932       1005006       -3.64%
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-12               355043        349566        -1.54%
BenchmarkDecoder_DecodeAll/alice29.txt.zst-12                462460        448584        -3.00%
BenchmarkDecoder_DecodeAll/html_x_4.zst-12                   236942        178127        -24.82%
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-12             19407         18649         -3.91%
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-12             8641          8690          +0.57%
BenchmarkDecoder_DecodeAll/urls.10K.zst-12                   1152539       1132885       -1.71%
BenchmarkDecoder_DecodeAll/html.zst-12                       99663         95678         -4.00%
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-12              8361          8438          +0.92%
BenchmarkDecoder_DecodeAllParallel/kppkn.gtb.zst-12          71902         68621         -4.56%
BenchmarkDecoder_DecodeAllParallel/geo.protodata.zst-12      14927         14423         -3.38%
BenchmarkDecoder_DecodeAllParallel/plrabn12.txt.zst-12       241417        229017        -5.14%
BenchmarkDecoder_DecodeAllParallel/lcet10.txt.zst-12         178932        169150        -5.47%
BenchmarkDecoder_DecodeAllParallel/asyoulik.txt.zst-12       59359         56624         -4.61%
BenchmarkDecoder_DecodeAllParallel/alice29.txt.zst-12        77809         74909         -3.73%
BenchmarkDecoder_DecodeAllParallel/html_x_4.zst-12           43376         32490         -25.10%
BenchmarkDecoder_DecodeAllParallel/paper-100k.pdf.zst-12     3572          3451          -3.39%
BenchmarkDecoder_DecodeAllParallel/fireworks.jpeg.zst-12     1535          1539          +0.26%
BenchmarkDecoder_DecodeAllParallel/urls.10K.zst-12           190123        185012        -2.69%
BenchmarkDecoder_DecodeAllParallel/html.zst-12               17203         16557         -3.76%
BenchmarkDecoder_DecodeAllParallel/comp-data.bin.zst-12      1509          1515          +0.40%

benchmark                                                    old MB/s     new MB/s     speedup
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-12               421.97       436.36       1.03x
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-12           1380.79      1392.31      1.01x
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-12            347.74       352.86       1.01x
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-12              406.06       418.48       1.03x
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-12            354.43       364.25       1.03x
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-12             325.60       337.59       1.04x
BenchmarkDecoder_DecoderSmall/html_x_4.zst-12                1721.62      2311.04      1.34x
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-12          5125.84      5307.43      1.04x
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-12          13123.74     13021.82     0.99x
BenchmarkDecoder_DecoderSmall/urls.10K.zst-12                596.18       620.02       1.04x
BenchmarkDecoder_DecoderSmall/html.zst-12                    1029.24      1061.94      1.03x
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-12           487.83       495.36       1.02x
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-12                  422.86       434.03       1.03x
BenchmarkDecoder_DecodeAll/geo.protodata.zst-12              1410.79      1309.62      0.93x
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-12               348.37       356.27       1.02x
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-12                 409.19       424.63       1.04x
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-12               352.57       358.10       1.02x
BenchmarkDecoder_DecodeAll/alice29.txt.zst-12                328.87       339.04       1.03x
BenchmarkDecoder_DecodeAll/html_x_4.zst-12                   1728.70      2299.49      1.33x
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-12             5276.43      5490.77      1.04x
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-12             14245.78     14165.08     0.99x
BenchmarkDecoder_DecodeAll/urls.10K.zst-12                   609.17       619.73       1.02x
BenchmarkDecoder_DecodeAll/html.zst-12                       1027.46      1070.26      1.04x
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-12              487.51       483.08       0.99x
BenchmarkDecoder_DecodeAllParallel/kppkn.gtb.zst-12          2563.50      2686.05      1.05x
BenchmarkDecoder_DecodeAllParallel/geo.protodata.zst-12      7944.49      8221.96      1.03x
BenchmarkDecoder_DecodeAllParallel/plrabn12.txt.zst-12       1995.97      2104.04      1.05x
BenchmarkDecoder_DecodeAllParallel/lcet10.txt.zst-12         2385.01      2522.94      1.06x
BenchmarkDecoder_DecodeAllParallel/asyoulik.txt.zst-12       2108.86      2210.70      1.05x
BenchmarkDecoder_DecodeAllParallel/alice29.txt.zst-12        1954.64      2030.32      1.04x
BenchmarkDecoder_DecodeAllParallel/html_x_4.zst-12           9443.01      12606.94     1.34x
BenchmarkDecoder_DecodeAllParallel/paper-100k.pdf.zst-12     28670.75     29674.17     1.03x
BenchmarkDecoder_DecodeAllParallel/fireworks.jpeg.zst-12     80178.39     79977.77     1.00x
BenchmarkDecoder_DecodeAllParallel/urls.10K.zst-12           3692.80      3794.81      1.03x
BenchmarkDecoder_DecodeAllParallel/html.zst-12               5952.37      6184.74      1.04x
BenchmarkDecoder_DecodeAllParallel/comp-data.bin.zst-12      2701.93      2691.29      1.00x

benchmark                                                    old allocs     new allocs     delta
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-12               1              1              +0.00%
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-12           1              1              +0.00%
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-12            1              1              +0.00%
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-12              1              1              +0.00%
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-12            1              1              +0.00%
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-12             1              1              +0.00%
BenchmarkDecoder_DecoderSmall/html_x_4.zst-12                1              1              +0.00%
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-12          1              1              +0.00%
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-12          1              1              +0.00%
BenchmarkDecoder_DecoderSmall/urls.10K.zst-12                1              1              +0.00%
BenchmarkDecoder_DecoderSmall/html.zst-12                    1              1              +0.00%
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-12           1              1              +0.00%
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-12                  0              0              +0.00%
BenchmarkDecoder_DecodeAll/geo.protodata.zst-12              0              0              +0.00%
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-12               0              0              +0.00%
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-12                 0              0              +0.00%
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-12               0              0              +0.00%
BenchmarkDecoder_DecodeAll/alice29.txt.zst-12                0              0              +0.00%
BenchmarkDecoder_DecodeAll/html_x_4.zst-12                   0              0              +0.00%
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-12             0              0              +0.00%
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-12             0              0              +0.00%
BenchmarkDecoder_DecodeAll/urls.10K.zst-12                   0              0              +0.00%
BenchmarkDecoder_DecodeAll/html.zst-12                       0              0              +0.00%
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-12              0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/kppkn.gtb.zst-12          0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/geo.protodata.zst-12      0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/plrabn12.txt.zst-12       0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/lcet10.txt.zst-12         0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/asyoulik.txt.zst-12       0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/alice29.txt.zst-12        0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/html_x_4.zst-12           0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/paper-100k.pdf.zst-12     0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/fireworks.jpeg.zst-12     0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/urls.10K.zst-12           0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/html.zst-12               0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/comp-data.bin.zst-12      0              0              +0.00%

benchmark                                                    old bytes     new bytes     delta
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-12               19459         19347         -0.58%
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-12           2704          2594          -4.07%
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-12            165860        158323        -4.54%
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-12              108902        105083        -3.51%
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-12            8997          8892          -1.17%
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-12             17441         16595         -4.85%
BenchmarkDecoder_DecoderSmall/html_x_4.zst-12                24948         18030         -27.73%
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-12          48            527           +997.92%
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-12          332           326           -1.81%
BenchmarkDecoder_DecoderSmall/urls.10K.zst-12                199639        195355        -2.15%
BenchmarkDecoder_DecoderSmall/html.zst-12                    48            48            +0.00%
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-12           48            48            +0.00%
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-12                  0             0             +0.00%
BenchmarkDecoder_DecodeAll/geo.protodata.zst-12              0             0             +0.00%
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-12               0             0             +0.00%
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-12                 0             0             +0.00%
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-12               0             0             +0.00%
BenchmarkDecoder_DecodeAll/alice29.txt.zst-12                0             0             +0.00%
BenchmarkDecoder_DecodeAll/html_x_4.zst-12                   0             0             +0.00%
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-12             0             0             +0.00%
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-12             0             0             +0.00%
BenchmarkDecoder_DecodeAll/urls.10K.zst-12                   0             0             +0.00%
BenchmarkDecoder_DecodeAll/html.zst-12                       0             0             +0.00%
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-12              0             0             +0.00%
BenchmarkDecoder_DecodeAllParallel/kppkn.gtb.zst-12          141           131           -7.09%
BenchmarkDecoder_DecodeAllParallel/geo.protodata.zst-12      18            18            +0.00%
BenchmarkDecoder_DecodeAllParallel/plrabn12.txt.zst-12       1302          1289          -1.00%
BenchmarkDecoder_DecodeAllParallel/lcet10.txt.zst-12         986           741           -24.85%
BenchmarkDecoder_DecodeAllParallel/asyoulik.txt.zst-12       78            75            -3.85%
BenchmarkDecoder_DecodeAllParallel/alice29.txt.zst-12        122           130           +6.56%
BenchmarkDecoder_DecodeAllParallel/html_x_4.zst-12           179           133           -25.70%
BenchmarkDecoder_DecodeAllParallel/paper-100k.pdf.zst-12     3             3             +0.00%
BenchmarkDecoder_DecodeAllParallel/fireworks.jpeg.zst-12     2             2             +0.00%
BenchmarkDecoder_DecodeAllParallel/urls.10K.zst-12           1539          1515          -1.56%
BenchmarkDecoder_DecodeAllParallel/html.zst-12               18            17            -5.56%
BenchmarkDecoder_DecodeAllParallel/comp-data.bin.zst-12      0             0             +0.00%
```

go1.17/arm64:
```
benchmark                                                   old ns/op     new ns/op     delta
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-4               7285584       7358412       +1.00%
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-4           1948236       1954437       +0.32%
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-4            23894495      23566187      -1.37%
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-4              17781669      17587395      -1.09%
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-4            6056783       5986820       -1.16%
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-4             7595705       7503978       -1.21%
BenchmarkDecoder_DecoderSmall/html_x_4.zst-4                4190851       4190728       -0.00%
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-4          553375        561100        +1.40%
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-4          348653        348737        +0.02%
BenchmarkDecoder_DecoderSmall/urls.10K.zst-4                21020394      20865234      -0.74%
BenchmarkDecoder_DecoderSmall/html.zst-4                    2038084       2009719       -1.39%
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-4           146524        156045        +6.50%
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-4                  900881        908068        +0.80%
BenchmarkDecoder_DecodeAll/geo.protodata.zst-4              240917        242009        +0.45%
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-4               2962741       2932156       -1.03%
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-4                 2203425       2174672       -1.30%
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-4               752544        743828        -1.16%
BenchmarkDecoder_DecodeAll/alice29.txt.zst-4                943281        931194        -1.28%
BenchmarkDecoder_DecodeAll/html_x_4.zst-4                   515398        517646        +0.44%
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-4             67341         68499         +1.72%
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-4             41159         41151         -0.02%
BenchmarkDecoder_DecodeAll/urls.10K.zst-4                   2594535       2577253       -0.67%
BenchmarkDecoder_DecodeAll/html.zst-4                       252830        249818        -1.19%
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-4              18349         19569         +6.65%
BenchmarkDecoder_DecodeAllParallel/kppkn.gtb.zst-4          226216        228030        +0.80%
BenchmarkDecoder_DecodeAllParallel/geo.protodata.zst-4      60384         60714         +0.55%
BenchmarkDecoder_DecodeAllParallel/plrabn12.txt.zst-4       748918        738658        -1.37%
BenchmarkDecoder_DecodeAllParallel/lcet10.txt.zst-4         555155        547405        -1.40%
BenchmarkDecoder_DecodeAllParallel/asyoulik.txt.zst-4       188566        186724        -0.98%
BenchmarkDecoder_DecodeAllParallel/alice29.txt.zst-4        236716        233740        -1.26%
BenchmarkDecoder_DecodeAllParallel/html_x_4.zst-4           129772        129420        -0.27%
BenchmarkDecoder_DecodeAllParallel/paper-100k.pdf.zst-4     16902         17164         +1.55%
BenchmarkDecoder_DecodeAllParallel/fireworks.jpeg.zst-4     10399         11602         +11.57%
BenchmarkDecoder_DecodeAllParallel/urls.10K.zst-4           652628        688743        +5.53%
BenchmarkDecoder_DecodeAllParallel/html.zst-4               63494         62539         -1.50%
BenchmarkDecoder_DecodeAllParallel/comp-data.bin.zst-4      4707          5002          +6.27%

benchmark                                                   old MB/s     new MB/s     speedup
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-4               202.39       200.39       0.99x
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-4           486.96       485.41       1.00x
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-4            161.33       163.58       1.01x
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-4              192.00       194.12       1.01x
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-4            165.34       167.27       1.01x
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-4             160.18       162.14       1.01x
BenchmarkDecoder_DecoderSmall/html_x_4.zst-4                781.89       781.92       1.00x
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-4          1480.37      1459.99      0.99x
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-4          2824.42      2823.74      1.00x
BenchmarkDecoder_DecoderSmall/urls.10K.zst-4                267.20       269.19       1.01x
BenchmarkDecoder_DecoderSmall/html.zst-4                    401.95       407.62       1.01x
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-4           222.54       208.97       0.94x
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-4                  204.60       202.98       0.99x
BenchmarkDecoder_DecodeAll/geo.protodata.zst-4              492.24       490.01       1.00x
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-4               162.64       164.34       1.01x
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-4                 193.68       196.24       1.01x
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-4               166.34       168.29       1.01x
BenchmarkDecoder_DecodeAll/alice29.txt.zst-4                161.23       163.33       1.01x
…