@klauspost 2022-01-09
Thanks!

This, AFAICT, copies the arrays (4K each) to the stack. 

Quite surprised that this actually is faster than referencing the table.