@lizthegrey 2022-01-09
> This, AFAICT, copies the arrays (4K each) to the stack.
> 
> Quite surprised that this actually is faster than referencing the table.

I was very surprised too. More investigation is going to be needed, heh. (I also did try passing pointers to the table, but that worked out slower for some reason)?