@klauspost 2022-01-09
"Preheating" lookup tables can have this effect.

Could you double check if you have the correct order in your compare? I am getting pretty much exactly the same, but in the opposite direction on a similar cpu (AMD Ryzen 9 3950X) :

```
λ benchcmp before.txt after2.txt
benchmark                                                    old ns/op     new ns/op     delta
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-32                  568428        587575        +3.37%
BenchmarkDecoder_DecodeAll/geo.protodata.zst-32              130568        132809        +1.72%
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-32               1812405       1862562       +2.77%
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-32                 1347352       1387654       +2.99%
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-32               468022        465707        -0.49%
BenchmarkDecoder_DecodeAll/alice29.txt.zst-32                609575        630171        +3.38%
BenchmarkDecoder_DecodeAll/html_x_4.zst-32                   248985        294998        +18.48%
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-32             23935         24671         +3.07%
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-32             9347          9368          +0.22%
BenchmarkDecoder_DecodeAll/urls.10K.zst-32                   1506474       1560133       +3.56%
BenchmarkDecoder_DecodeAll/html.zst-32                       140460        144998        +3.23%
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-32              11009         11026         +0.15%
BenchmarkDecoder_DecodeAllParallel/kppkn.gtb.zst-32          29158         29665         +1.74%
BenchmarkDecoder_DecodeAllParallel/geo.protodata.zst-32      6803          6942          +2.04%
BenchmarkDecoder_DecodeAllParallel/plrabn12.txt.zst-32       94311         97400         +3.28%
BenchmarkDecoder_DecodeAllParallel/lcet10.txt.zst-32         70224         72004         +2.53%
BenchmarkDecoder_DecodeAllParallel/asyoulik.txt.zst-32       24210         24355         +0.60%
BenchmarkDecoder_DecodeAllParallel/alice29.txt.zst-32        31580         32460         +2.79%
BenchmarkDecoder_DecodeAllParallel/html_x_4.zst-32           13222         17177         +29.91%
BenchmarkDecoder_DecodeAllParallel/paper-100k.pdf.zst-32     1359          1416          +4.19%
BenchmarkDecoder_DecodeAllParallel/fireworks.jpeg.zst-32     1961          1904          -2.91%
BenchmarkDecoder_DecodeAllParallel/urls.10K.zst-32           76334         77385         +1.38%
BenchmarkDecoder_DecodeAllParallel/html.zst-32               7505          7590          +1.13%
BenchmarkDecoder_DecodeAllParallel/comp-data.bin.zst-32      923           987           +6.96%
```