@lizthegrey 2022-01-09
```
$ go test -bench=Decode -run=None > /tmp/array_1.18
$ git checkout master
$ go test -bench=Decode -run=None > /tmp/master_1.18
$ ~/hny/go/bin/benchcmp /tmp/master_1.18 /tmp/array_1.18 
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                                    old ns/op     new ns/op     delta
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-12               3423669       3368844       -1.60%
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-12           670911        680979        +1.50%
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-12            11570774      10931823      -5.52%
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-12              8364803       8139978       -2.69%
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-12            2808659       2755397       -1.90%
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-12             3695290       3598235       -2.63%
BenchmarkDecoder_DecoderSmall/html_x_4.zst-12                1912378       1435330       -24.95%
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-12          158371        155247        -1.97%
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-12          75512         75103         -0.54%
BenchmarkDecoder_DecoderSmall/urls.10K.zst-12                9169653       9240729       +0.78%
BenchmarkDecoder_DecoderSmall/html.zst-12                    799481        773395        -3.26%
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-12           64981         65682         +1.08%
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-12                  432947        421317        -2.69%
BenchmarkDecoder_DecodeAll/geo.protodata.zst-12              83696         83171         -0.63%
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-12               1379241       1356588       -1.64%
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-12                 1040574       1012633       -2.69%
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-12               354805        349062        -1.62%
BenchmarkDecoder_DecodeAll/alice29.txt.zst-12                459306        447757        -2.51%
BenchmarkDecoder_DecodeAll/html_x_4.zst-12                   238385        176887        -25.80%
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-12             19248         18625         -3.24%
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-12             8569          8784          +2.51%
BenchmarkDecoder_DecodeAll/urls.10K.zst-12                   1152896       1139355       -1.17%
BenchmarkDecoder_DecodeAll/html.zst-12                       100290        95572         -4.70%
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-12              8217          8547          +4.02%
BenchmarkDecoder_DecodeAllParallel/kppkn.gtb.zst-12          70846         69131         -2.42%
BenchmarkDecoder_DecodeAllParallel/geo.protodata.zst-12      14641         14624         -0.12%
BenchmarkDecoder_DecodeAllParallel/plrabn12.txt.zst-12       235646        231551        -1.74%
BenchmarkDecoder_DecodeAllParallel/lcet10.txt.zst-12         176418        172322        -2.32%
BenchmarkDecoder_DecodeAllParallel/asyoulik.txt.zst-12       63351         57562         -9.14%
BenchmarkDecoder_DecodeAllParallel/alice29.txt.zst-12        77801         75349         -3.15%
BenchmarkDecoder_DecodeAllParallel/html_x_4.zst-12           43818         32565         -25.68%
BenchmarkDecoder_DecodeAllParallel/paper-100k.pdf.zst-12     3534          3471          -1.78%
BenchmarkDecoder_DecodeAllParallel/fireworks.jpeg.zst-12     1584          1558          -1.64%
BenchmarkDecoder_DecodeAllParallel/urls.10K.zst-12           189539        184289        -2.77%
BenchmarkDecoder_DecodeAllParallel/html.zst-12               16993         16967         -0.15%
BenchmarkDecoder_DecodeAllParallel/comp-data.bin.zst-12      1489          1516          +1.81%

benchmark                                                    old MB/s     new MB/s     speedup
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-12               430.70       437.71       1.02x
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-12           1414.05      1393.15      0.99x
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-12            333.16       352.63       1.06x
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-12              408.14       419.42       1.03x
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-12            356.55       363.44       1.02x
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-12             329.26       338.14       1.03x
BenchmarkDecoder_DecoderSmall/html_x_4.zst-12                1713.47      2282.96      1.33x
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-12          5172.67      5276.74      1.02x
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-12          13040.83     13111.98     1.01x
BenchmarkDecoder_DecoderSmall/urls.10K.zst-12                612.53       607.82       0.99x
BenchmarkDecoder_DecoderSmall/html.zst-12                    1024.67      1059.23      1.03x
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-12           501.81       496.46       0.99x
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-12                  425.73       437.49       1.03x
BenchmarkDecoder_DecodeAll/geo.protodata.zst-12              1416.89      1425.84      1.01x
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-12               349.37       355.20       1.02x
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-12                 410.11       421.43       1.03x
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-12               352.81       358.62       1.02x
BenchmarkDecoder_DecodeAll/alice29.txt.zst-12                331.13       339.67       1.03x
BenchmarkDecoder_DecodeAll/html_x_4.zst-12                   1718.23      2315.60      1.35x
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-12             5319.94      5498.00      1.03x
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-12             14364.88     14013.03     0.98x
BenchmarkDecoder_DecodeAll/urls.10K.zst-12                   608.98       616.21       1.01x
BenchmarkDecoder_DecodeAll/html.zst-12                       1021.04      1071.45      1.05x
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-12              496.07       476.89       0.96x
BenchmarkDecoder_DecodeAllParallel/kppkn.gtb.zst-12          2601.69      2666.25      1.02x
BenchmarkDecoder_DecodeAllParallel/geo.protodata.zst-12      8099.79      8108.88      1.00x
BenchmarkDecoder_DecodeAllParallel/plrabn12.txt.zst-12       2044.85      2081.01      1.02x
BenchmarkDecoder_DecodeAllParallel/lcet10.txt.zst-12         2419.00      2476.50      1.02x
BenchmarkDecoder_DecodeAllParallel/asyoulik.txt.zst-12       1975.97      2174.66      1.10x
BenchmarkDecoder_DecodeAllParallel/alice29.txt.zst-12        1954.85      2018.46      1.03x
BenchmarkDecoder_DecodeAllParallel/html_x_4.zst-12           9347.79      12578.00     1.35x
BenchmarkDecoder_DecodeAllParallel/paper-100k.pdf.zst-12     28978.20     29497.48     1.02x
BenchmarkDecoder_DecodeAllParallel/fireworks.jpeg.zst-12     77720.77     78991.28     1.02x
BenchmarkDecoder_DecodeAllParallel/urls.10K.zst-12           3704.17      3809.70      1.03x
BenchmarkDecoder_DecodeAllParallel/html.zst-12               6026.12      6035.39      1.00x
BenchmarkDecoder_DecodeAllParallel/comp-data.bin.zst-12      2737.24      2689.11      0.98x

benchmark                                                    old allocs     new allocs     delta
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-12               1              1              +0.00%
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-12           1              1              +0.00%
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-12            1              1              +0.00%
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-12              1              1              +0.00%
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-12            1              1              +0.00%
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-12             1              1              +0.00%
BenchmarkDecoder_DecoderSmall/html_x_4.zst-12                1              1              +0.00%
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-12          1              1              +0.00%
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-12          1              1              +0.00%
BenchmarkDecoder_DecoderSmall/urls.10K.zst-12                1              1              +0.00%
BenchmarkDecoder_DecoderSmall/html.zst-12                    1              1              +0.00%
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-12           1              1              +0.00%
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-12                  0              0              +0.00%
BenchmarkDecoder_DecodeAll/geo.protodata.zst-12              0              0              +0.00%
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-12               0              0              +0.00%
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-12                 0              0              +0.00%
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-12               0              0              +0.00%
BenchmarkDecoder_DecodeAll/alice29.txt.zst-12                0              0              +0.00%
BenchmarkDecoder_DecodeAll/html_x_4.zst-12                   0              0              +0.00%
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-12             0              0              +0.00%
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-12             0              0              +0.00%
BenchmarkDecoder_DecodeAll/urls.10K.zst-12                   0              0              +0.00%
BenchmarkDecoder_DecodeAll/html.zst-12                       0              0              +0.00%
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-12              0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/kppkn.gtb.zst-12          0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/geo.protodata.zst-12      0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/plrabn12.txt.zst-12       0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/lcet10.txt.zst-12         0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/asyoulik.txt.zst-12       0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/alice29.txt.zst-12        0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/html_x_4.zst-12           0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/paper-100k.pdf.zst-12     0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/fireworks.jpeg.zst-12     0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/urls.10K.zst-12           0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/html.zst-12               0              0              +0.00%
BenchmarkDecoder_DecodeAllParallel/comp-data.bin.zst-12      0              0              +0.00%

benchmark                                                    old bytes     new bytes     delta
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-12               19573         19236         -1.72%
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-12           2633          2528          -3.99%
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-12            165860        160198        -3.41%
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-12              108744        107411        -1.23%
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-12            8937          8851          -0.96%
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-12             17385         16745         -3.68%
BenchmarkDecoder_DecoderSmall/html_x_4.zst-12                24173         18262         -24.45%
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-12          535           48            -91.03%
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-12          327           331           +1.22%
BenchmarkDecoder_DecoderSmall/urls.10K.zst-12                196990        195074        -0.97%
BenchmarkDecoder_DecoderSmall/html.zst-12                    48            48            +0.00%
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-12           48            48            +0.00%
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-12                  0             0             +0.00%
BenchmarkDecoder_DecodeAll/geo.protodata.zst-12              0             0             +0.00%
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-12               0             0             +0.00%
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-12                 0             0             +0.00%
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-12               0             0             +0.00%
BenchmarkDecoder_DecodeAll/alice29.txt.zst-12                0             0             +0.00%
BenchmarkDecoder_DecodeAll/html_x_4.zst-12                   0             0             +0.00%
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-12             0             0             +0.00%
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-12             0             0             +0.00%
BenchmarkDecoder_DecodeAll/urls.10K.zst-12                   0             0             +0.00%
BenchmarkDecoder_DecodeAll/html.zst-12                       0             0             +0.00%
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-12              0             0             +0.00%
BenchmarkDecoder_DecodeAllParallel/kppkn.gtb.zst-12          136           135           -0.74%
BenchmarkDecoder_DecodeAllParallel/geo.protodata.zst-12      18            18            +0.00%
BenchmarkDecoder_DecodeAllParallel/plrabn12.txt.zst-12       1306          1299          -0.54%
BenchmarkDecoder_DecodeAllParallel/lcet10.txt.zst-12         949           891           -6.11%
BenchmarkDecoder_DecodeAllParallel/asyoulik.txt.zst-12       79            75            -5.06%
BenchmarkDecoder_DecodeAllParallel/alice29.txt.zst-12        127           130           +2.36%
BenchmarkDecoder_DecodeAllParallel/html_x_4.zst-12           182           133           -26.92%
BenchmarkDecoder_DecodeAllParallel/paper-100k.pdf.zst-12     3             3             +0.00%
BenchmarkDecoder_DecodeAllParallel/fireworks.jpeg.zst-12     2             2             +0.00%
BenchmarkDecoder_DecodeAllParallel/urls.10K.zst-12           1506          1452          -3.59%
BenchmarkDecoder_DecodeAllParallel/html.zst-12               18            17            -5.56%
BenchmarkDecoder_DecodeAllParallel/comp-data.bin.zst-12      0             0             +0.00%
```